pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "event-system"
version = "1.0.0"
description = "Production-ready event flow platform with FastAPI and message brokers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.32",
  "asyncpg>=0.29.0",
  "alembic>=1.13.2",
  "redis>=5.0.8",
  "aio-pika>=9.4.0",
  "aiokafka>=0.11.0",
  "nats-py>=2.7.2",
  "aiogram>=3.11.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "prometheus-fastapi-instrumentator>=7.0.0",
  "prometheus-client>=0.20.0",
  "httpx>=0.27.0",
  "structlog>=24.4.0",
  "tenacity>=9.0.0",
  "pybreaker>=1.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.8",
  "pytest-cov>=5.0.0",
  "mypy>=1.11.1",
  "black>=24.8.0",
  "isort>=5.13.2",
  "flake8>=7.1.1",
  "bandit>=1.7.9",
  "safety>=3.2.7",
  "locust>=2.31.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
addopts = "-q -m 'not integration'"
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
  "integration: marks tests requiring real infrastructure",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
plugins = []

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
